requirements.txt
langchain
langgraph
langchain_community 
langchainhub
langchain-groq 
langchain-huggingface
chromadb
langchain
openai
qdrant-client
tavily-python
streamlit
sentence-transformersstreamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv